+++ lib/fitting.rb
@@ -1,10 +1,9 @@
 require 'fitting/version'
 require 'fitting/configuration'
-require 'fitting/storage/documentation'
 require 'fitting/matchers/response_matcher'
-require 'fitting/statistics'
 require 'fitting/documentation'
 require 'fitting/storage/responses'
+require 'fitting/railtie' if defined?(Rails)
 
 module Fitting
   class << self
@@ -13,7 +12,7 @@ module Fitting
     end
 
     def configuration
-      @configuration ||= Configuration.new
+      @configuration ||= Configuration.craft
     end
 
     def statistics
@@ -21,14 +20,44 @@ module Fitting
 
       RSpec.configure do |config|
         config.after(:each, type: :controller) do
-          responses.add(response)
+          responses.add(response, self.inspect)
         end
 
         config.after(:suite) do
           responses.statistics.save
-          responses.statistics.cover_save if Fitting.configuration.json_schema_cover
         end
       end
+    end
+
+    def save_test_data
+      responses = Fitting::Storage::Responses.new
+
+      RSpec.configure do |config|
+        config.after(:each, type: :controller) do
+          responses.add(response, self.inspect)
+        end
+
+        config.after(:suite) do
+          responses.tests.save
+        end
+      end
+    end
+  end
+
+  def self.loaded_tasks=(val)
+    @loaded_tasks = val
+  end
+
+  def self.loaded_tasks
+    @loaded_tasks
+  end
+
+  def self.load_tasks
+    return if loaded_tasks
+    self.loaded_tasks = true
+
+    Dir[File.join(File.dirname(__FILE__), 'tasks', '**/*.rake')].each do |rake|
+      load rake
     end
   end
 end