+++ lib/fitting/storage/responses.rb
@@ -1,41 +1,24 @@
-require 'fitting/storage/white_list'
+require 'fitting/statistics'
+require 'fitting/tests'
+require 'fitting/records/tested/request'
 
 module Fitting
   module Storage
     class Responses
       def initialize
-        @responses = []
+        @tested_requests = []
       end
 
-      def add(response)
-        @responses.push(
-          Fitting::Response.new(
-            response,
-            Fitting::Storage::Documentation.tomogram
-          )
-        )
+      def add(env_response, test_title = '')
+        @tested_requests.push(Fitting::Records::Tested::Request.new(env_response, test_title))
       end
 
       def statistics
-        @white_list = white_list
-        Fitting::Statistics.new(
-          Fitting::Documentation.new(
-            Fitting::Storage::Documentation.tomogram,
-            @white_list
-          ),
-          @responses.uniq,
-          Fitting.configuration.strict
-        )
+        Fitting::Statistics.new(@tested_requests)
       end
 
-      private
-
-      def white_list
-        Fitting::Storage::WhiteList.new(
-          Fitting.configuration.white_list,
-          Fitting.configuration.resource_white_list,
-          Fitting::Storage::Documentation.tomogram.to_resources
-        ).to_a
+      def tests
+        Fitting::Tests.new(@tested_requests)
       end
     end
   end