+++ spec/lib/fitting_spec.rb
@@ -16,8 +16,12 @@ RSpec.describe Fitting do
     it 'does not raise exception' do
       allow(RSpec).to receive(:configure).and_yield(config)
       allow(config).to receive(:after).and_yield
-      allow(described_class).to receive(:response)
-      allow(Fitting::Storage::Responses).to receive(:new).and_return(responses)
+      allow(config).to receive(:before).and_yield
+      allow(described_class).to receive(:response).and_return(double(request: nil))
+      allow(Fitting::Storage::Responses).to receive(:new).and_return(double(
+                                                                       add: nil,
+                                                                       statistics: double(save: nil)
+      ))
       expect { subject.statistics }.not_to raise_exception
     end
   end